Extract process.go from main.go
